remove task state, add dataflow & worker state

+++ src/WebJobs.Script/Dispatch/LanguageWorkerChannel.cs
@@ -2,151 +2,118 @@
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.IO;
 using System.Linq;
 using System.Reactive.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Azure.WebJobs.Host;
+using System.Threading.Tasks.Dataflow;
 using Microsoft.Azure.WebJobs.Script.Abstractions.Rpc;
 using Microsoft.Azure.WebJobs.Script.Description;
 using Microsoft.Azure.WebJobs.Script.Description.Script;
 using Microsoft.Azure.WebJobs.Script.Eventing;
+using Microsoft.Azure.WebJobs.Script.Eventing.Rpc;
 using Microsoft.Azure.WebJobs.Script.Grpc.Messages;
 using Microsoft.Azure.WebJobs.Script.Rpc;
 using Microsoft.Extensions.Logging;
 
+using MsgType = Microsoft.Azure.WebJobs.Script.Grpc.Messages.StreamingMessage.ContentOneofCase;
+
 namespace Microsoft.Azure.WebJobs.Script.Dispatch
 {
-    // TODO: move to RPC project?
     internal class LanguageWorkerChannel : ILanguageWorkerChannel
     {
-        private readonly TimeSpan timeoutStart = TimeSpan.FromSeconds(60);
-        private readonly TimeSpan timeoutInit = TimeSpan.FromSeconds(60);
-        private readonly TimeSpan timeoutLoad = TimeSpan.FromSeconds(60);
-        private readonly TimeSpan timeoutInvoke = TimeSpan.FromMinutes(60);
+        private readonly TimeSpan timeoutStart = TimeSpan.FromSeconds(20);
+        private readonly TimeSpan timeoutInit = TimeSpan.FromSeconds(20);
         private readonly ScriptHostConfiguration _scriptConfig;
         private readonly IScriptEventManager _eventManager;
         private readonly IWorkerProcessFactory _processFactory;
+        private readonly IObservable<FunctionRegistrationContext> _functionRegistrations;
         private readonly WorkerConfig _workerConfig;
         private readonly Uri _serverUri;
         private readonly ILogger _logger;
+        private string _workerId;
         private Process _process;
-        private IDictionary<FunctionMetadata, Task<FunctionLoadResponse>> _functionLoadState = new Dictionary<FunctionMetadata, Task<FunctionLoadResponse>>();
+        private IDictionary<string, BufferBlock<ScriptInvocationContext>> _functionInputBuffers = new Dictionary<string, BufferBlock<ScriptInvocationContext>>();
+        private ConcurrentDictionary<string, ScriptInvocationContext> _executingInvocations = new ConcurrentDictionary<string, ScriptInvocationContext>();
+        private IObservable<InboundEvent> _inboundWorkerEvents;
 
-        private AsyncLazy<WorkerInfo> _workerInfo;
+        private List<IDisposable> _inputLinks = new List<IDisposable>();
+        private List<IDisposable> _eventSubscriptions = new List<IDisposable>();
 
-        public LanguageWorkerChannel(ScriptHostConfiguration scriptConfig, IScriptEventManager eventManager, IWorkerProcessFactory processFactory, WorkerConfig workerConfig, Uri serverUri, ILogger logger)
+        private bool disposedValue;
+
+        public LanguageWorkerChannel(
+            ScriptHostConfiguration scriptConfig,
+            IScriptEventManager eventManager,
+            IWorkerProcessFactory processFactory,
+            IObservable<FunctionRegistrationContext> functionRegistrations,
+            WorkerConfig workerConfig,
+            Uri serverUri,
+            ILogger logger)
         {
             _scriptConfig = scriptConfig;
             _eventManager = eventManager;
             _processFactory = processFactory;
+            _functionRegistrations = functionRegistrations;
             _workerConfig = workerConfig;
             _serverUri = serverUri;
             _logger = logger;
-            _workerInfo = new AsyncLazy<WorkerInfo>((ct) => StartWorkerAsync(), new CancellationTokenSource());
+            _workerId = Guid.NewGuid().ToString();
+
+            _inboundWorkerEvents = _eventManager.OfType<InboundEvent>()
+                .Where(msg => msg.WorkerId == _workerId);
+
+            StartWorker();
         }
 
-        private async Task SendAsync(StreamingMessage msg, string workerId = null)
-        {
-            workerId = workerId ?? (await _workerInfo).Id;
-            var msgEvent = new RpcEvent(workerId, msg);
-            _eventManager.Publish(msgEvent);
-        }
+        public string Id => _workerId;
 
-        private async Task<StreamingMessage> ReceiveResponseAsync(string responseId, TimeSpan timeout)
-        {
-            TaskCompletionSource<StreamingMessage> tcs = new TaskCompletionSource<StreamingMessage>();
-            IDisposable subscription = null;
-            subscription = _eventManager
-                .OfType<RpcEvent>()
-                .Where(msg => msg.Origin == RpcEvent.MessageOrigin.Worker && msg.Message?.RequestId == responseId)
-                .Timeout(timeout)
-                .Subscribe(response =>
-                {
-                    tcs.SetResult(response.Message);
-                    subscription?.Dispose();
-                }, exc =>
-                {
-                    tcs.SetException(exc);
-                    subscription?.Dispose();
-                });
-            return await tcs.Task;
-        }
+        public WorkerConfig Config => _workerConfig;
 
-        private async Task<StreamingMessage> ReceiveResponseAsync(StreamingMessage request, TimeSpan timeout, string workerId = null)
+        public void Invoke(ScriptInvocationContext context)
         {
-            request.RequestId = Guid.NewGuid().ToString();
-            var receiveTask = ReceiveResponseAsync(request.RequestId, timeout);
-            await SendAsync(request, workerId);
-            return await receiveTask;
-        }
+            if (context.CancellationToken.IsCancellationRequested)
+            {
+                context.ResultSource.SetCanceled();
+                return;
+            }
 
-        public async Task<ScriptInvocationResult> InvokeAsync(FunctionMetadata functionMetadata, ScriptInvocationContext context)
-        {
-            FunctionLoadResponse loadResponse = await _functionLoadState.GetOrAdd(functionMetadata, metadata => LoadAsync(metadata));
+            var functionMetadata = context.FunctionMetadata;
 
             InvocationRequest invocationRequest = new InvocationRequest()
             {
-                FunctionId = loadResponse.FunctionId,
+                FunctionId = functionMetadata.FunctionId,
                 InvocationId = context.ExecutionContext.InvocationId.ToString(),
             };
-
             foreach (var pair in context.BindingData)
             {
-                invocationRequest.TriggerMetadata.Add(pair.Key, await pair.Value.ToRpcAsync().ConfigureAwait(false));
+                invocationRequest.TriggerMetadata.Add(pair.Key, pair.Value.ToRpc());
             }
-
             foreach (var input in context.Inputs)
             {
                 invocationRequest.InputData.Add(new ParameterBinding()
                 {
                     Name = input.name,
-
-                    // TODO: get rid of this async. We can't handle async reading of streams (http body)
-                    Data = await input.val.ToRpcAsync().ConfigureAwait(false)
+                    Data = input.val.ToRpc()
                 });
             }
 
-            var response = await ReceiveResponseAsync(new StreamingMessage()
+            _executingInvocations.TryAdd(invocationRequest.InvocationId, context);
+
+            Send(new StreamingMessage
             {
                 InvocationRequest = invocationRequest
-            }, timeoutInvoke);
-            InvocationResponse invocationResponse = response.InvocationResponse;
-            invocationResponse.Result.VerifySuccess();
-
-            IDictionary<string, object> bindingsDictionary = invocationResponse.OutputData
-                    .ToDictionary(binding => binding.Name, binding => binding.Data.ToObject());
-
-            return new ScriptInvocationResult()
-            {
-                Outputs = bindingsDictionary,
-                Return = invocationResponse?.ReturnValue?.ToObject()
-            };
-        }
-
-        public async Task HandleFileEventAsync(FileSystemEventArgs fileEvent)
-        {
-            FileChangeEventRequest request = new FileChangeEventRequest()
-            {
-                FullPath = fileEvent.FullPath,
-                Name = fileEvent.Name,
-                Type = (FileChangeEventRequest.Types.Type)fileEvent.ChangeType
-            };
-
-            await SendAsync(new StreamingMessage()
-            {
-                FileChangeEventRequest = request,
             });
         }
 
-        private async Task<FunctionLoadResponse> LoadAsync(FunctionMetadata metadata)
+        public void Register(FunctionRegistrationContext context)
         {
+            FunctionMetadata metadata = context.Metadata;
+            _functionInputBuffers[context.Metadata.FunctionId] = context.InputBuffer;
             FunctionLoadRequest request = new FunctionLoadRequest()
             {
-                FunctionId = Guid.NewGuid().ToString(),
+                FunctionId = metadata.FunctionId,
                 Metadata = new RpcFunctionMetadata()
                 {
                     Name = metadata.Name,
@@ -165,63 +132,118 @@ namespace Microsoft.Azure.WebJobs.Script.Dispatch
                 });
             }
 
-            var response = await ReceiveResponseAsync(new StreamingMessage()
+            Send(new StreamingMessage
             {
                 FunctionLoadRequest = request
-            }, timeoutLoad);
-
-            response.FunctionLoadResponse.Result.VerifySuccess();
-            return response.FunctionLoadResponse;
+            });
         }
 
-        public void Register(FunctionMetadata functionMetadata)
+        internal void InitWorker(RpcEvent startEvent)
         {
-            _functionLoadState[functionMetadata] = LoadAsync(functionMetadata);
-        }
+            _inboundWorkerEvents.Where(msg => msg.MessageType == MsgType.WorkerInitResponse)
+                .Timeout(timeoutInit)
+                .Take(1)
+                .Subscribe(WorkerReady, HandleWorkerError);
 
-        internal async Task<WorkerInfo> StartWorkerAsync()
-        {
-            await StopAsync();
-
-            string workerId = Guid.NewGuid().ToString();
-            string requestId = Guid.NewGuid().ToString();
-
-            try
+            Send(new StreamingMessage
             {
-                var startStreamTask = ReceiveResponseAsync(requestId, timeoutStart);
-
-                var workerContext = new WorkerCreateContext()
+                WorkerInitRequest = new WorkerInitRequest()
                 {
-                    RequestId = requestId,
-                    WorkerId = workerId,
-                    WorkerConfig = _workerConfig,
-                    WorkingDirectory = _scriptConfig.RootScriptPath,
-                    Logger = _logger,
-                    ServerUri = _serverUri,
+                    HostVersion = ScriptHost.Version
+                }
+            });
+        }
+
+        internal void HandleWorkerError(Exception exc)
+        {
+            // unlink function inputs
+            foreach (var link in _inputLinks)
+            {
+                link.Dispose();
+            }
+            _logger.LogError($"Worker {_workerId} encountered an error.", exc);
+            _eventManager.Publish(new WorkerErrorEvent(this, exc));
+        }
+
+        internal void WorkerReady(RpcEvent initEvent)
+        {
+            var initMessage = initEvent.Message.WorkerInitResponse;
+            if (initMessage.Result.IsFailure(out Exception exc))
+            {
+                HandleWorkerError(exc);
+                return;
+            }
+
+            _eventSubscriptions.Add(_functionRegistrations.Subscribe(Register));
+
+            _eventSubscriptions.Add(_inboundWorkerEvents.Where(msg => msg.MessageType == MsgType.FunctionLoadResponse)
+                .Subscribe((msg) => LoadResponse(msg.Message.FunctionLoadResponse)));
+
+            _eventSubscriptions.Add(_inboundWorkerEvents.Where(msg => msg.MessageType == MsgType.InvocationResponse)
+                .Subscribe((msg) => InvokeResponse(msg.Message.InvocationResponse)));
+
+            _eventManager.Publish(new WorkerReadyEvent
+            {
+                Id = _workerId,
+                Version = initMessage.WorkerVersion,
+                Capabilities = initMessage.Capabilities,
+                Config = _workerConfig,
+            });
+        }
+
+        internal void InvokeResponse(InvocationResponse invokeResponse)
+        {
+            if (_executingInvocations.TryRemove(invokeResponse.InvocationId, out ScriptInvocationContext context)
+                && invokeResponse.Result.IsSuccess(context.ResultSource))
+            {
+                IDictionary<string, object> bindingsDictionary = invokeResponse.OutputData
+                    .ToDictionary(binding => binding.Name, binding => binding.Data.ToObject());
+
+                var result = new ScriptInvocationResult()
+                {
+                    Outputs = bindingsDictionary,
+                    Return = invokeResponse?.ReturnValue?.ToObject()
                 };
-
-                _process = _processFactory.CreateWorkerProcess(workerContext);
-                StartProcess(workerId, _process);
-                await startStreamTask;
-
-                StreamingMessage response = await ReceiveResponseAsync(new StreamingMessage()
-                {
-                    WorkerInitRequest = new WorkerInitRequest()
-                    {
-                        HostVersion = ScriptHost.Version
-                    }
-                }, timeoutInit, workerId);
-
-                WorkerInitResponse initResponse = response.WorkerInitResponse;
-                initResponse.Result.VerifySuccess();
-
-                return new WorkerInfo(workerId, initResponse.WorkerVersion, initResponse.Capabilities);
+                context.ResultSource.SetResult(result);
             }
-            catch (Exception exc)
+        }
+
+        internal void LoadResponse(FunctionLoadResponse loadResponse)
+        {
+            if (loadResponse.Result.IsFailure(out Exception e))
             {
-                _logger.LogError($"Worker {workerId} was unable to start", exc);
-                throw;
+                _logger.LogError($"Function {loadResponse.FunctionId} failed to load", e);
+
+                // load retry?
             }
+            else
+            {
+                var inputBuffer = _functionInputBuffers[loadResponse.FunctionId];
+                var invokeBlock = new ActionBlock<ScriptInvocationContext>(ctx => Invoke(ctx));
+                var disposableLink = inputBuffer.LinkTo(invokeBlock);
+                _inputLinks.Add(disposableLink);
+            }
+        }
+
+        internal void StartWorker()
+        {
+            _inboundWorkerEvents.Where(msg => msg.MessageType == MsgType.StartStream)
+                .Timeout(timeoutStart)
+                .Take(1)
+                .Subscribe(InitWorker, HandleWorkerError);
+
+            var workerContext = new WorkerCreateContext()
+            {
+                RequestId = Guid.NewGuid().ToString(),
+                WorkerId = _workerId,
+                WorkerConfig = _workerConfig,
+                WorkingDirectory = _scriptConfig.RootScriptPath,
+                Logger = _logger,
+                ServerUri = _serverUri,
+            };
+
+            _process = _processFactory.CreateWorkerProcess(workerContext);
+            StartProcess(_workerId, _process);
         }
 
         internal void StartProcess(string workerId, Process process)
@@ -237,13 +259,12 @@ namespace Microsoft.Azure.WebJobs.Script.Dispatch
             process.EnableRaisingEvents = true;
             process.Exited += (s, e) =>
             {
-                if (process.ExitCode > 0)
+                if (process.ExitCode != 0)
                 {
-                    _logger.LogError($"Worker {workerId} pid {process.Id} exited with code {process.ExitCode}");
+                    HandleWorkerError(new Exception($"Worker process with pid {process.Id} exited with code {process.ExitCode}"));
                 }
                 process.WaitForExit();
                 process.Close();
-                process.Dispose();
             };
 
             process.Start();
@@ -251,20 +272,31 @@ namespace Microsoft.Azure.WebJobs.Script.Dispatch
             process.BeginOutputReadLine();
         }
 
-        public Task StopAsync()
+        private void Send(StreamingMessage msg)
         {
-            // TODO: send cancellation warning
-            // TODO: Close request stream for each worker pool
-            _process?.Kill();
-            _process = null;
-            return Task.CompletedTask;
+            _eventManager.Publish(new OutboundEvent(_workerId, msg));
+        }
+
+        protected virtual void Dispose(bool disposing)
+        {
+            if (!disposedValue)
+            {
+                if (disposing)
+                {
+                    _process?.Kill();
+                    _process?.Dispose();
+                    foreach (var sub in _eventSubscriptions)
+                    {
+                        sub.Dispose();
+                    }
+                }
+                disposedValue = true;
+            }
         }
 
         public void Dispose()
         {
-            _workerInfo.Dispose();
-            _process?.Kill();
-            _process?.Dispose();
+            Dispose(true);
         }
     }
 }
